test: fix tests
