--- packages/react/jest.config.js
@@ -1,10 +1,0 @@
-export default {
-  preset: 'ts-jest',
-  testEnvironment: 'jsdom',
-  collectCoverageFrom: ['./src/**'],
-  globals: {
-    'ts-jest': {
-      tsconfig: 'tsconfig.test.json',
-    },
-  },
-};
