+++ packages/react/package.json
@@ -27,8 +27,8 @@
     "build": "pnpm prebuild && cross-env NODE_ENV=production rollup -c",
     "dev": "rollup -cw",
     "prepublishOnly": "pnpm build && pnpm test",
-    "test": "jest",
-    "test:ci": "jest --ci --coverage"
+    "test": "uvu -r tsm -r global-jsdom/register tests -i common",
+    "test:ci": "nyc pnpm test"
   },
   "files": [
     "dist"
@@ -44,14 +44,12 @@
   },
   "devDependencies": {
     "@felte/common": "workspace:*",
-    "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^12.1.2",
     "@testing-library/react-hooks": "^7.0.2",
     "@types/react": "^17.0.37",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-test-renderer": "^17.0.2",
-    "ts-jest": "^26.5.0"
+    "react-test-renderer": "^17.0.2"
   },
   "exports": {
     ".": {
