+++ packages/validator-zod/package.json
@@ -35,6 +35,7 @@
   },
   "devDependencies": {
     "felte": "workspace:*",
+    "svelte": "^3.46.4",
     "zod": "^3.2.0"
   },
   "peerDependencies": {
